requirements.txt
torch==2.2.2
numpy
matplotlib
pillow